Revert "Merge pull request #366 from dbt-labs/manual-revert-to-working"
This reverts commit 2dc6b74f2026742450cec9a600a97a64818144b8, reversing
changes made to 86aab285690600287b1c6c37614cbbb713da2e22.

+++ pkg/dbt_cloud/webhook.go
@@ -32,7 +32,7 @@ type WebhookWrite struct {
 	Description string   `json:"description,omitempty"`
 	ClientUrl   string   `json:"client_url"`
 	EventTypes  []string `json:"event_types,omitempty"`
-	JobIds      []int    `json:"job_ids"`
+	JobIds      []int64  `json:"job_ids"`
 	Active      bool     `json:"active,omitempty"`
 }
 
@@ -71,7 +71,7 @@ func (c *Client) CreateWebhook(
 	description string,
 	clientUrl string,
 	eventTypes []string,
-	jobIds []int,
+	jobIds []int64,
 	active bool,
 ) (*WebhookRead, error) {
